Fix two crashes in TouchImageView.java
If the drawable is set to null during a double-tap zoom or other interpolated event, we can get crashes. This can happen if, for example, Glide is used to load a new image into the TouchImageView in the middle of a zoom. This change fixes crashes that come inside the functions the new if() statements are guarding, both of which I've experienced.